requirements.txt
aniso8601==8.0.0
anytree==2.8.0
asgiref==3.2.10
astroid==2.3.3
autopep8==1.4.4
beautifulsoup4==4.8.2
braintree==4.0.0
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
crayons==0.3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.41.0
django-filter==2.2.0
django-six==1.0.4
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
ffmpeg==1.4
ffmpeg-python==0.2.0
firebase-admin==4.0.0
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
future==0.18.2
gcloud==0.17.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
graphviz==0.13.2
grpcio==1.27.2
httplib2==0.17.0
idna==2.10
imageio==2.8.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
jws==0.1.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
msgpack==1.0.0
mutagen==1.44.0
mysql==0.0.2
mysql-connector-python==8.0.18
mysqlclient==1.4.6
networkx==2.4
numpy==1.19.2
oauth==1.0.1
oauth2client==3.0.0
oauthlib==3.1.0
opencv-python==4.4.0.44
paho-mqtt==1.5.1
pandas==0.25.3
pbr==5.4.3
piexif==1.1.3
Pillow==7.2.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pydub==0.23.1
pygame==1.9.6
pyglet==1.5.2
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-firebase==1.2
python-jwt==2.0.1
python-vlc==3.0.7110
python3-openid==3.1.0
pytz==2020.1
PyWavelets==1.1.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
selenium==3.141.0
simpleaudio==1.0.4
six==1.15.0
sklearn==0.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
SQLAlchemy==1.3.18
sqlparse==0.3.1
stegano==0.9.8
stepic==0.5.0
stevedore==1.31.0
stripe==2.43.0
ttkthemes==3.0.0
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.3.3
